[Transforms] refactor agg value extractors into their own class
